requirements.txt
google-generativeai
python-dotenv
streamlit
openai-whisper